feat(user): 优化用户服务层实现
1. 重构用户模块接口定义，将变量形式的接口定义改为类型定义
2. 优化用户密码安全性，增加密码盐值和加盐哈希存储
3. 新增用户状态批量更新接口，提高更新效率
4. 完善日志记录，增强系统可观测性

+++ internal/common/response.go
@@ -12,10 +12,10 @@ func New(code int, msg string, data any) *Response {
 }
 
 type Page struct {
-	List  any `json:"list"`
-	Total int `json:"total"`
+	List  any   `json:"list"`
+	Total int64 `json:"total"`
 }
 
-func NewPage(list any, total int) *Page {
+func NewPage(list any, total int64) *Page {
 	return &Page{List: list, Total: total}
 }
